+++ db/migrate/20160316181024_change_type_of_compiled_content_on_notifications.rb
@@ -1,0 +1,5 @@
+class ChangeTypeOfCompiledContentOnNotifications < ActiveRecord::Migration
+  def change
+    change_column :notifications, :compiled_content, :text
+  end
+end