+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20151106132950) do
+ActiveRecord::Schema.define(version: 20160316181024) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -39,7 +39,7 @@ ActiveRecord::Schema.define(version: 20151106132950) do
     t.string   "category"
     t.json     "additions"
     t.string   "nkey"
-    t.string   "compiled_content"
+    t.text     "compiled_content"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.integer  "contact_id"