+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - :buib: Updated API documentation
 - :hammer: Fixed `ZazoActivity` score criteria (calculate all friends, raised ratio)
 - :hammer: Fixed `Contact::GetSuggestions` service to return only last non-friend contacts
+- :hammer: Changed database type of `notifications.compiled_content` from `string` to `text` 
 - :hammer: Refactored code
 
 ### v0.3.0